feat(a11y): update to new spec #137, #165
This commit will alter the scope of a11y
support for Auro icons. All descriptive
text needed for a11y is handled by the
<title> attribute and <desc> is optional.

As of this commit, icons are not hidden
from screen readers by default. This feature
is left up to the user of the icons per install.

Changes to be committed:
modified:   .svglintrc.js
modified:   scripts/generate.js
modified:   src/data/icons.json