perf: update SVGs to support new token name spec #211
This commit reverts the revert for using the new Auro design token name
spec. The revert happened due to a side-effect with auro-icon and the
customSize feature. The auro-icon element was updated to allow for more
flexibility in token use.

See https://github.com/AlaskaAirlines/auro-icon/pull/95

This update includes a solution that applies a variable with a variable
fallback with a hard value fallback. The --auro token was specifically
listed first as this is an order of expectation from in-the-wild auro
elements using the legacy --auro token.

It has been verified that regardless of version of tokens a
consumer has installed, or none at all, the customSize feature will work
as expected.

Changes to be committed:
modified:   src/_icons.scss
modified:   src/data/icons.json