perf: remove unnecessary style property
This commit removed an unnecessary `height` CSS property that was
deprecated with the new addition of the CSS variable fallback on the
next line.

Leaving this property in required the support of all the legacy design
tokens for this one processed value.

Changes to be committed:
modified:   src/_icons.scss