perf: add support for CSS parts
This commit adds support for CSS Parts on the SVG allowing for targetd
styling of the SVG within the element as needed.

This update was added to the SVGO script versus the DOM building script
due to SVGO's intention to strip away non-uniform attributes and no part
of their API to support retaining this attribute.

Changes to be committed:
modified:   scripts/generate.js