build: update to support BETA releases
https://github.com/orgs/AlaskaAirlines/discussions/556

This code diff represents changes made to multiple files in a Git repository.
It includes modifications to the `.github/settings.yml` file, the
`.github/workflows/testPublish.yml` file, and the `package.json` file.

In the `.github/settings.yml` file, the changes include:
- Enabling the requirement for code owner reviews before merging pull requests.
- Adding dismissal restrictions for pull request reviews, allowing only
the user "blackfalcon" to dismiss reviews.
- Enabling the prevention of merge commits on matching branches.

In the `.github/workflows/testPublish.yml` file, the change includes:
- Modifying the condition for triggering the release job. It will now trigger
on push events to either the `master` branch or the `beta` branch.

In the `package.json` file, the change includes:
- Modifying the `release` configuration to include both the `master` and `beta`
branches for semantic release. The `beta` branch is marked as a prerelease branch.

Changes to be committed:
modified:   .github/settings.yml
modified:   .github/workflows/testPublish.yml
modified:   package.json